+++ lib/ruby-next/language.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-gem "parser", "~> 2.6.3.105"
-gem "unparser", "~> 0.4.5"
+gem "parser", ">= 2.7.0.0"
+gem "unparser", ">= 0.4.7"
 
 require "set"